refactor: Change course package
