--- src/main/java/com/hexagonal/courseregistration/course/AuthorityException.java
@@ -1,9 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-import com.hexagonal.courseregistration.common.application.ForbiddenException;
-
-public class AuthorityException extends ForbiddenException {
-  public AuthorityException(Message message) {
-    super(message.detail());
-  }
-}
