--- src/main/java/com/hexagonal/courseregistration/course/CheckExistCoursePort.java
@@ -1,5 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-public interface CheckExistCoursePort {
-  boolean check(String courseName);
-}
