--- src/main/java/com/hexagonal/courseregistration/course/CheckProfessorPort.java
@@ -1,5 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-public interface CheckProfessorPort {
-  boolean check(Long id);
-}
