--- src/main/java/com/hexagonal/courseregistration/course/CourseException.java
@@ -1,9 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-import com.hexagonal.courseregistration.common.application.DomainException;
-
-public class CourseException extends DomainException {
-  public CourseException(Message message) {
-    super(message.detail());
-  }
-}
