--- src/main/java/com/hexagonal/courseregistration/course/CourseTime.java
@@ -1,7 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-record CourseTime(
-  DayOfWeek dayOfWeek,
-  Duration duration
-) {
-}
