--- src/main/java/com/hexagonal/courseregistration/course/DayOfWeek.java
@@ -1,9 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-enum DayOfWeek {
-  MONDAY,
-  TUESDAY,
-  WEDNESDAY,
-  THURSDAY,
-  FRIDAY
-}
