--- src/main/java/com/hexagonal/courseregistration/course/Message.java
@@ -1,15 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Message {
-  NOT_PROFESSOR("Not Professor"),
-  NOT_EXIST_USER("Not Exist User"),
-  EXIST_COURSE("Exist Course Name"),
-  INVALID_SCORE("Invalid Score"),
-  INVALID_DURATION("Invalid Course Duration");
-  private final String detail;
-}
