--- src/main/java/com/hexagonal/courseregistration/course/RegisterRequest.java
@@ -1,10 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-record RegisterRequest(
-  Long userId,
-  String courseName,
-  Score score,
-  Integer capacity,
-  CourseTime courseTime
-) {
-}
