--- src/main/java/com/hexagonal/courseregistration/course/SaveCoursePort.java
@@ -1,5 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-public interface SaveCoursePort {
-  void save(RegisterRequest request);
-}
