--- src/main/java/com/hexagonal/courseregistration/course/UserException.java
@@ -1,9 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-import com.hexagonal.courseregistration.common.application.DomainException;
-
-public class UserException extends DomainException {
-  public UserException(Message message) {
-    super(message.detail());
-  }
-}
