+++ src/main/java/com/hexagonal/courseregistration/course/application/domain/CourseTime.java
@@ -1,0 +1,7 @@
+package com.hexagonal.courseregistration.course.application.domain;
+
+public record CourseTime(
+  DayOfWeek dayOfWeek,
+  Duration duration
+) {
+}
