+++ src/main/java/com/hexagonal/courseregistration/course/application/domain/DayOfWeek.java
@@ -1,0 +1,9 @@
+package com.hexagonal.courseregistration.course.application.domain;
+
+public enum DayOfWeek {
+  MONDAY,
+  TUESDAY,
+  WEDNESDAY,
+  THURSDAY,
+  FRIDAY
+}
