+++ src/main/java/com/hexagonal/courseregistration/course/application/error/AuthorityException.java
@@ -1,0 +1,9 @@
+package com.hexagonal.courseregistration.course.application.error;
+
+import com.hexagonal.courseregistration.common.application.ForbiddenException;
+
+public class AuthorityException extends ForbiddenException {
+  public AuthorityException(Message message) {
+    super(message.detail());
+  }
+}
