+++ src/main/java/com/hexagonal/courseregistration/course/application/port/CheckExistCoursePort.java
@@ -1,0 +1,5 @@
+package com.hexagonal.courseregistration.course.application.port;
+
+public interface CheckExistCoursePort {
+  boolean check(String courseName);
+}
