+++ src/main/java/com/hexagonal/courseregistration/course/application/port/CheckProfessorPort.java
@@ -1,0 +1,5 @@
+package com.hexagonal.courseregistration.course.application.port;
+
+public interface CheckProfessorPort {
+  boolean check(Long id);
+}
