+++ src/main/java/com/hexagonal/courseregistration/course/application/port/SaveCoursePort.java
@@ -1,0 +1,5 @@
+package com.hexagonal.courseregistration.course.application.port;
+
+public interface SaveCoursePort {
+  void save(RegisterRequest request);
+}
