+++ src/main/java/com/hexagonal/courseregistration/user/application/CheckExistUserPort.java
@@ -1,7 +1,5 @@
 package com.hexagonal.courseregistration.user.application;
 
-import com.hexagonal.courseregistration.user.application.Authority;
-
 public interface CheckExistUserPort {
   boolean check(String idNumber, Authority authority);
 }
