+++ src/main/java/com/hexagonal/courseregistration/user/application/SaveUserPort.java
@@ -1,7 +1,5 @@
 package com.hexagonal.courseregistration.user.application;
 
-import com.hexagonal.courseregistration.user.application.User;
-
 public interface SaveUserPort {
   void save(User user);
 }
