--- src/test/java/com/hexagonal/courseregistration/course/ScoreTest.java
@@ -1,17 +1,0 @@
-package com.hexagonal.courseregistration.course;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-class ScoreTest {
-  @Test
-  void constructor_smaller_than_1() {
-    assertThrows(CourseException.class, () -> new Score(0));
-  }
-
-  @Test
-  void constructor_bigger_than_3() {
-    assertThrows(CourseException.class, () -> new Score(4));
-  }
-}
